\name{addc}
\alias{addc}
\title{ Adds a column vector to a transformed design matrix and updates its
        QR-decomposition}
\description{
 See Marazzi A. (1993), p.355}
\usage{
addc(x, n = nrow(x), l, j, ip)
}
\arguments{
  \item{x}{ See reference}
  \item{n}{ See reference}
  \item{l}{ See reference}
  \item{j}{ See reference}
  \item{ip}{ See reference}
}

\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.355
}
\keyword{utilities}

