\name{airef0}
\alias{airef0}
\title{ Asymptotic relative efficiency of a general M-estimate for a model
        with mu quantitative covariates with or without a constant term} 
                       
\description{
 See Marazzi A. (1993), p.167}
\usage{
airef0(expsi = psi, exu = ucv, exw = www, itype = .dFvGet()$ite, mu, 
       ialfa = .dFvGet()$ial, sigmx = 1, upper = .dFvGet()$upr, 
       til = .dFvGet()$tli, maxit = .dFvGet()$mxe, tol = .dFvGet()$tlo)
}
\arguments{
  \item{expsi}{ See reference}
  \item{exu}{ See reference}
  \item{exw}{ See reference}
  \item{itype}{ See reference}
  \item{mu}{ See reference}
  \item{ialfa}{ See reference}
  \item{sigmx}{ See reference}
  \item{upper}{ See reference}
  \item{til}{ See reference}
  \item{maxit}{ See reference}
  \item{tol}{ See reference}
}
\value{
  See reference}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.167
}
\keyword{robust}

