\name{binprd}
\alias{binprd}
\title{ Binomial probability distribution }
\description{
 See Marazzi A. (1993), p.367}
\usage{
binprd(k, n, p)
}
\arguments{
  \item{k}{ See reference}
  \item{n}{ See reference}
  \item{p}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.367
}
\keyword{distribution}

