\name{cerf}
\alias{cerf}
\title{ Complemented error function (single precision)}
\description{
 See Marazzi A. (1993), p.380}
\usage{
cerf(x)
}
\arguments{
  \item{x}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.380
}
\keyword{distribution}

