\name{cfrcov}
\alias{cfrcov}
\title{Computation of fC.fC.inv(AT A) for a given matrix A and scale factor fC}
\description{
 See Marazzi A. (1993), p.242}
\usage{
cfrcov(a, nvar, fc, tau = .dFvGet()$tua)
}
\arguments{
  \item{a}{ See reference}
  \item{nvar}{ See reference}
  \item{fc}{ See reference}
  \item{tau}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.242
}
\keyword{utilities}

