\name{chi}
\alias{chi}
\title{Chi weight function for location and regression }
\description{
 See Marazzi A. (1993), p.322}
\usage{
chi(s)
}

\arguments{
  \item{s}{ A scalar input value }
}

\value{
 The value of the chi function for s
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.322
}
\keyword{robust}

