\name{Chi}
\alias{Chi}
\title{Chi weight function for location and regression}
\description{
 See Marazzi A. (1993), p.322}
\usage{
Chi(svals)
}
\arguments{
  \item{svals}{ A vector of input values }
}
\value{
 The values of the chi function for each element of svals
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.32
}
\keyword{robust}

