\name{chisq}
\alias{chisq}
\title{Cumulative Chi-square distribution function}
\description{
 See Marazzi A. (1993), p.373}
\usage{
chisq(kode = 1, ifn, x)
}
\arguments{
  \item{kode}{ See reference}
  \item{ifn}{ See reference}
  \item{x}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.373
}
\keyword{distribution}

