\name{cibeat}
\alias{cibeat}
\title{Determination of the parameter d of the Huber weight function }
\description{
 See Marazzi A. (1993), p.247}
\usage{
cibeat(a2 = .dFvGet()$aa2, b2 = .dFvGet()$bb2, nvar)
}
\arguments{
  \item{a2}{ See reference}
  \item{b2}{ See reference}
  \item{nvar}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.247
}
\keyword{robust}

