\name{cicloc}
\alias{cicloc}
\title{ Determination of the parameter c of the Huber weight function from the
        proportion eps of contamination}
\description{
 See Marazzi A. (1993), p.243}
\usage{
cicloc(eps = .dFvGet()$esp, tol = .dFvGet()$tlo)
}
\arguments{
  \item{eps}{ See reference}
  \item{tol}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.243
}
\keyword{robust}

