\name{comval}
\alias{comval}
\title{Gives the current values ofthe parameters of the ROBETH 
       subroutine common blocks}
\description{
 See Marazzi A. (1993), p.405}
\usage{
comval()
}

\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.405
}
\keyword{utilities}

