\name{cquant}
\alias{cquant}
\title{ Inverse of the cumulative Chi2-distribution function }
\description{
 See Marazzi A. (1993), p.374}
\usage{
cquant(p, ifn, tol = 5e-06, maxit = 50)
}
\arguments{
  \item{p}{ See reference}
  \item{ifn}{ See reference}
  \item{tol}{ See reference}
  \item{maxit}{ See reference}
}

\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.374
}
\keyword{distribution}

