\name{Dbinom}
\alias{Dbinom}
\title{ Diagonal matrix D for the binomial case in discrete GLM }
\description{
 See Marazzi A. (1993), p.310}
\usage{
Dbinom(y, ci, vtheta, wa, ni, f0, oi = 0, kap = 1e-06)
}
\arguments{
  \item{y}{ See reference}
  \item{ci}{ See reference}
  \item{vtheta}{ See reference}
  \item{wa}{ See reference}
  \item{ni}{ See reference}
  \item{f0}{ See reference}
  \item{oi}{ See reference}
  \item{kap}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.310
}
\keyword{robust}

