\name{dfrpar}
\alias{dfrpar}
\title{ Sets default parameters for regression estimates }
\description{
See Marazzi A. (1993), p.398 and p.406}
\usage{
dfrpar(x, etype, upar = -1, psipar = -1)
}
\arguments{
  \item{x}{ See reference}
  \item{etype}{ See reference}
  \item{upar}{ See reference}
  \item{psipar}{ See reference}
}

\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California.  p.398 and p.406
}
\keyword{utilities}

