\name{dfvals}
\alias{dfvals}
\title{ Provide default values for most scalar parameters used by the Robeth subroutines}
\description{
 See Marazzi A. (1993), p.404}
\usage{
dfvals()
}

\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.404
}
\keyword{utilities}

