\name{.dFvGet}
\alias{.dFvGet}
\title{ Provide the current values of the scalar parameters used by the Robeth subroutines}
\description{
 See Marazzi A. (1993), p.404}
\usage{
.dFvGet()
}

\value{
  A list with the current values of the scalar parameters uaed in robeth. See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.404
}
\keyword{utilities}

