\name{.dFvPut}
\alias{.dFvPut}
\title{ Set the current values of the scalar parameters used by the Robeth subroutines}
\description{
 See Marazzi A. (1993), p.404}
\usage{
.dFvPut(vals, nams)
}
\arguments{
  \item{vals}{The vector of parameter values}
  \item{nams}{ The vector of parameter names to be modified}
}

\value{
  No value returned. The named parameters are internaly modified.
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.404
}
\keyword{utilities}