\name{dotp}
\alias{dotp}
\title{ Forms the scalar (dot) product of two vectors }
\description{
 See Marazzi A. (1993), p.350}
\usage{
dotp(x, y, n = nrow(x), incx = 1, incy = 1)
}
\arguments{
  \item{x}{ See reference}
  \item{y}{ See reference}
  \item{n}{ See reference}
  \item{incx}{ See reference}
  \item{incy}{ See reference}
}

\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.350
}
\keyword{utilities}

