\name{exch}
\alias{exch}
\title{ Exchanges two columns of a symmetric matrix }
\description{
 See Marazzi A. (1993), p.364}
\usage{
exch(s, n, h, k)
}
\arguments{
  \item{s}{ See reference}
  \item{n}{ See reference}
  \item{h}{ See reference}
  \item{k}{ See reference}
}

\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.364
}
\keyword{utilities}

