\name{fcum}
\alias{fcum}
\title{ Cumulative F-distribution function }
\description{
 See Marazzi A. (1993), p.379}
\usage{
fcum(n1, n2, x)
}
\arguments{
  \item{n1}{ See reference}
  \item{n2}{ See reference}
  \item{x}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.37
}
\keyword{distribution}

