\name{Fn.Exp.f}
\alias{Fn.Exp.f}
\title{ Parametric estimate of survival cdf}
\description{
  Parametric estimate of survival cdf}
\usage{
 Fn.Exp.f(z, y , delta, mu, sigma, lambda, zero=1e-4)
}
\arguments{
  \item{z}{ See reference}
  \item{y}{ See reference}
  \item{delta}{ See reference}
  \item{mu}{ See reference}
  \item{sigma}{ See reference}
  \item{lambda}{ See reference}
  \item{zero}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (2010) Robust estimation of the extended log-gamma (not yet published)
}
\keyword{robust}



