\name{fstord}
\alias{fstord}
\title{ Determination of the j-th order statistic }
\description{
 See Marazzi A. (1993), p.389}
\usage{
fstord(y, j)
}
\arguments{
  \item{y}{ See reference}
  \item{j}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California.  p.389
}
\keyword{utilities}

