\name{gaussd}
\alias{gaussd}
\title{ Cumulative Gaussian distribution function (double precision) }
\description{
 See Marazzi A. (1993), p.371}
\usage{
gaussd(kode = 1, x)
}
\arguments{
  \item{kode}{ See reference}
  \item{x}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.371
}
\keyword{distribution}

