\name{gfedca}
\alias{gfedca}
\title{ Diagonal matrices D_G and E_G}
\description{
 See Marazzi A. (1993), p.309}
\usage{
gfedca(vtheta, ci, wa, ni, oi = 0, icase = .dFvGet()$ics)
}
\arguments{
  \item{vtheta}{ See reference}
  \item{ci}{ See reference}
  \item{wa}{ See reference}
  \item{ni}{ See reference}
  \item{oi}{ See reference}
  \item{icase}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.309
}
\keyword{utilities}

