\name{h12}
\alias{h12}
\title{ Constructs and/or applies a single elementary Householder transformation }
\description{
 See Marazzi A. (1993), p.359}

\usage{
h12(mode, lpivot, l1, u, up, c, ice, icv, ncv)
}
\arguments{
  \item{mode}{ See reference}
  \item{lpivot}{ See reference}
  \item{l1}{ See reference}
  \item{u}{ See reference}
  \item{up}{ See reference}
  \item{c}{ See reference}
  \item{ice}{ See reference}
  \item{icv}{ See reference}
  \item{ncv}{ See reference}
}

\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.359
}
\keyword{utilities}

