\name{ingama}
\alias{ingama}
\title{ Incomplete Gamma-integral function }
\description{
 See Marazzi A. (1993), p.381}

\usage{
ingama(x, p)
}
\arguments{
  \item{x}{ See reference}
  \item{p}{ See reference}
}

\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.381
}
\keyword{distribution}

