\name{kfascv}
\alias{kfascv}
\title{ Backtransformation of the covariance matrix of the 
        coefficient estimates }
\description{
 See Marazzi A. (1993), p.152}
\usage{
kfascv(xt, cov, k = np, mdx = nrow(xt), f = .dFvGet()$fff, sg, ip)
}
\arguments{
  \item{xt}{ See reference}
  \item{cov}{ See reference}
  \item{k}{ See reference}
  \item{mdx}{ See reference}
  \item{f}{ See reference}
  \item{sg}{ See reference}
  \item{ip}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.152
}
\keyword{robust}

