\name{kfedcb}
\alias{kfedcb}
\title{Diagonal hat matrices D_M, E_M, D_S, and E_S }
\description{
 See Marazzi A. (1993), p.159}
\usage{
kfedcb(wgt, rs, expsi = psi, expsp = psp, sigma, itype = .dFvGet()$ite)
}
\arguments{
  \item{wgt}{ See reference}
  \item{rs}{ See reference}
  \item{expsi}{ See reference}
  \item{expsp}{ See reference}
  \item{sigma}{ See reference}
  \item{itype}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.15
}
\keyword{robust}

