\name{kffacv}
\alias{kffacv}
\title{ Correction factor f_H for the covariance matrix of a
        Huber-type estimate }
\description{
 See Marazzi A. (1993), p.154}
\usage{
kffacv(rs, expsi = psi, expsp = psp, np, sigma)
}
\arguments{
  \item{rs}{ See reference}
  \item{expsi}{ See reference}
  \item{expsp}{ See reference}
  \item{np}{ See reference}
  \item{sigma}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.154
}
\keyword{robust}

