\name{kiedch}
\alias{kiedch}
\title{ Diagonal matrices D_M, E_M, D_S, E_S when psi
        is the Huber function }
\description{
 See Marazzi A. (1993), p.156}
\usage{
kiedch(wgt, c = .dFvGet()$ccc, itype = .dFvGet()$ite)
}
\arguments{
  \item{wgt}{ See reference}
  \item{c}{ See reference}
  \item{itype}{ See reference}
}

\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.156
}
\keyword{robust}

