\name{kiedcu}
\alias{kiedcu}
\title{Diagonal matrices D_M, E_M, D_S, E_S when psi
       is a user-supplied function }
\description{
 See Marazzi A. (1993), p.157}
\usage{
kiedcu(wgt, expsi = psi, itype = .dFvGet()$ite, upper = .dFvGet()$upr, 
       til = .dFvGet()$tli)
}
\arguments{
  \item{wgt}{ See reference}
  \item{expsi}{ See reference}
  \item{itype}{ See reference}
  \item{upper}{ See reference}
  \item{til}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.157
}
\keyword{robust}

