\name{lgama}
\alias{lgama}
\title{ Logarithm at the Gamma-function at the point x }
\description{
 See Marazzi A. (1993), p.383}
\usage{
lgama(x)
}
\arguments{
  \item{x}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.383
}
\keyword{utilities}

