\name{libet0}
\alias{libet0}
\title{ Computation of Beta0 = Phi_inv(0.75)}
\description{
 See Marazzi A. (1993), p.46}
\usage{
libet0()
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.46
}
\keyword{robust}

