\name{libeth}
\alias{libeth}
\title{Computation of Int Chi(s) dPhi(s) when  Chi=Psi.Psi/2 and Psi is 
        the Huber function }
\description{
 See Marazzi A. (1993), p.44}
\usage{
libeth(d = .dFvGet()$ddd)
}
\arguments{
  \item{d}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.44
}
\keyword{robust}

