\name{libetu}
\alias{libetu}
\title{ Computation of Int Chi(s) dPhi(s) when  Chi is a user-supplied function }
\description{
 See Marazzi A. (1993), p.45}
\usage{
libetu(exchi = chi, upper = .dFvGet()$upr, til = .dFvGet()$tli)
}
\arguments{
  \item{exchi}{ See reference}
  \item{upper}{ See reference}
  \item{til}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.45
}
\keyword{robust}

