\name{liclls}
\alias{liclls}
\title{ Classical estimates of mean and standard deviation }
\description{
 See Marazzi A. (1993), p.27}
\usage{
liclls(y)
}
\arguments{
  \item{y}{ Vector of observations }
}

\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.27
}
\keyword{utilities}

