\name{liepsh}
\alias{liepsh}
\title{ Computation of Int Psi(s).Psi(s) dPhi(s) and Int Psi'(s) dPhi(s) when  Psi is 
        the Huber function  }
\description{
 See Marazzi A. (1993), p.47}
\usage{
liepsh(c = .dFvGet()$ccc)
}
\arguments{
  \item{c}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.47
}
\keyword{robust}

