\name{liepsu}
\alias{liepsu}
\title{ Computation of Int Psi(s).Psi(s) dPhi(s) and Int Psi'(s) dPhi(s) when  Psi is 
        a user-supplied external function }
\description{
 See Marazzi A. (1993), p.48}
\usage{
liepsu(expsi = psi, upper = .dFvGet()$upr, til = .dFvGet()$tli)
}
\arguments{
  \item{expsi}{ See reference}
  \item{upper}{ See reference}
  \item{til}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.48
}
\keyword{robust}

