\name{liindh}
\alias{liindh}
\title{Inverts the approximate null distribution of the
       one-sample Wilcoxon test statistic }
\description{
 See Marazzi A. (1993), p.36}
\usage{
liindh(alpha = .dFvGet()$alf, n)
}
\arguments{
  \item{alpha}{ See reference}
  \item{n}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.36
}
\keyword{utilities}

