\name{liinds}
\alias{liinds}
\title{ Inverts the approximate null distribution of the sign test statistic }
\description{
 See Marazzi A. (1993), p.35}
\usage{
liinds(alpha = .dFvGet()$alf, n)
}
\arguments{
  \item{alpha}{ See reference}
  \item{n}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.35
}
\keyword{distribution}

