\name{liindw}
\alias{liindw}
\title{ Inverts the approximate null distribution of the
        Mann-Whitney test statistic}
\description{
 See Marazzi A. (1993), p.43}
\usage{
liindw(alpha = .dFvGet()$alf, m, n)
}
\arguments{
  \item{alpha}{ See reference}
  \item{m}{ See reference}
  \item{n}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.43
}
\keyword{utilities}

