\name{lilars}
\alias{lilars}
\title{Median an median absolute deviation}
\description{
 See Marazzi A. (1993), p.28}
\usage{
lilars(y, isort = .dFvGet()$isr)
}
\arguments{
  \item{y}{ See reference}
  \item{isort}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.28
}
\keyword{utilities}

