\name{littst}
\alias{littst}
\title{ t-test for the shift parameter }
\description{
 See Marazzi A. (1993), p.37}
\usage{
littst(x, y, alpha = .dFvGet()$alf)
}
\arguments{
  \item{x}{ See reference}
  \item{y}{ See reference}
  \item{alpha}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.37
}
\keyword{utilities}

