\name{lmdd}
\alias{lmdd}
\title{ Median and median absolute deviation }
\description{
 See Marazzi A. (1993), p.388}
\usage{
lmdd(x, isort = 1)
}
\arguments{
  \item{x}{ See reference}
  \item{isort}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.388
}
\keyword{utilities}

