\name{lyhdle}
\alias{lyhdle}
\title{ Hodges-Lehman estimate and confidence intervals for the 
        center of symmetry based on the one-sample Wilcoxon test}
\description{
 See Marazzi A. (1993), p.33}
\usage{
lyhdle(y, isort = .dFvGet()$isr, k, tol = .dFvGet()$tlo, 
       maxit = .dFvGet()$mxt)
}
\arguments{
  \item{y}{ See reference}
  \item{isort}{ See reference}
  \item{k}{ See reference}
  \item{tol}{ See reference}
  \item{maxit}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.33
}
\keyword{utilities}

