\name{lymnwt}
\alias{lymnwt}
\title{ Nonparametric estimate and confidence intervals for the
        shift parameter based on the Mann-Whitney test statistic}
\description{
 See Marazzi A. (1993), p.41}

\usage{
lymnwt(x, y, isort = .dFvGet()$isr, k, tol = .dFvGet()$tlo, 
       maxit = .dFvGet()$mxt)
}
\arguments{
  \item{x}{ See reference}
  \item{y}{ See reference}
  \item{isort}{ See reference}
  \item{k}{ See reference}
  \item{tol}{ See reference}
  \item{maxit}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.41
}
\keyword{utilities}

