\name{lytau2}
\alias{lytau2}
\title{tau-test for the shift parameter }
\description{
 See Marazzi A. (1993), p.38}
\usage{
lytau2(z, expsi = psi, expsp = psp, exchi = chi, exrho = rho, m, n, 
       tol = .dFvGet()$tlo, gam = .dFvGet()$gma, isigma = .dFvGet()$isg, 
       maxit = .dFvGet()$mxt, nitmon = .dFvGet()$ntm)
}
\arguments{
  \item{z}{ See reference}
  \item{expsi}{ See reference}
  \item{expsp}{ See reference}
  \item{exchi}{ See reference}
  \item{exrho}{ See reference}
  \item{m}{ See reference}
  \item{n}{ See reference}
  \item{tol}{ See reference}
  \item{gam}{ See reference}
  \item{isigma}{ See reference}
  \item{maxit}{ See reference}
  \item{nitmon}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.38
}
\keyword{robust}

