\name{mchl}
\alias{mchl}
\title{Cholesky decomposition of a symmetric matrix}
\description{
 See Marazzi A. (1993), p.353}
\usage{
mchl(a, n)
}
\arguments{
  \item{a}{ See reference}
  \item{n}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.353
}
\keyword{algebra}
