\name{mchld}
\alias{mchld}
\title{ Cholesky decomposition of a symmetric matrix (double precision) }
\description{
 See Marazzi A. (1993), p.353}
\usage{
mchld(a, n)
}
\arguments{
  \item{a}{ See reference}
  \item{n}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.353
}
\keyword{algebra}
