\name{mffd}
\alias{mffd}
\title{ Multiplies a full matrix by a full matrix (double precision) }
\description{
 See Marazzi A. (1993), p.339}
\usage{
mffd(a, b, m = nrow(a))
}
\arguments{
  \item{a}{ See reference}
  \item{b}{ See reference}
  \item{m}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.339
}
\keyword{utilities}
