\name{mfragr}
\alias{mfragr}
\title{ Generation and comparison of all regressions on subsets
        of covariates}
\description{
 See Marazzi A. (1993), p.258}
\usage{
mfragr(x, y, vp, nc, itype = .dFvGet()$ith, c = .dFvGet()$ccc, 
       tol = .dFvGet()$tlo, gam = .dFvGet()$gma, 
       maxit = .dFvGet()$mxt, sigmac, sigmar)
}
\arguments{
  \item{x}{ See reference}
  \item{y}{ See reference}
  \item{vp}{ See reference}
  \item{nc}{ See reference}
  \item{itype}{ See reference}
  \item{c}{ See reference}
  \item{tol}{ See reference}
  \item{gam}{ See reference}
  \item{maxit}{ See reference}
  \item{sigmac}{ See reference}
  \item{sigmar}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.258
}
\keyword{robust}

