\name{mfyd}
\alias{mfyd}
\title{ Multiplies a full matrix by a vector (double precision) }
\description{
 See Marazzi A. (1993), p.342}
\usage{
mfyd(a, y, m = nrow(a), iye = 1, ize = 1)
}
\arguments{
  \item{a}{ See reference}
  \item{y}{ See reference}
  \item{m}{ See reference}
  \item{iye}{ See reference}
  \item{ize}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.342
}
\keyword{utilities}

