\name{mhat}
\alias{mhat}
\title{ Computes the diagonal elements of the hat matrix }
\description{
 See Marazzi A. (1993), p.354}
\usage{
mhat(x, n = nrow(x), k = np, sh)
}
\arguments{
  \item{x}{ See reference}
  \item{n}{ See reference}
  \item{k}{ See reference}
  \item{sh}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.354
}
\keyword{algebra}

