\name{minv}
\alias{minv}
\title{ Inverts a triangular matrix }
\description{
 See Marazzi A. (1993), p.348}
\usage{
minv(r, n, tau = .dFvGet()$tua)
}
\arguments{
  \item{r}{ See reference}
  \item{n}{ See reference}
  \item{tau}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.348
}
\keyword{utilities}

