\name{mlyd}
\alias{mlyd}
\title{ Multiplies a lower-triangular matrix by a vector (double precision) }
\description{
 See Marazzi A. (1993), p.346}

\usage{
mlyd(a, y, n, iye = 1)
}
\arguments{
  \item{a}{ See reference}
  \item{y}{ See reference}
  \item{n}{ See reference}
  \item{iye}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.346
}
\keyword{utilities}

