\name{msf}
\alias{msf}
\title{ Multiplies a symmetric matrix by a full matrix }
\description{
 See Marazzi A. (1993), p.340}
\usage{
msf(a, b, n)
}
\arguments{
  \item{a}{ See reference}
  \item{b}{ See reference}
  \item{n}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.340
}
\keyword{utilities}

