\name{msf1d}
\alias{msf1d}
\title{ Multiplies a symmetric matrix by a full matrix when the result
        is a symmetric matrix }
\description{
 See Marazzi A. (1993), p.341}
\usage{
msf1d(a, b)
}
\arguments{
  \item{a}{ See reference}
  \item{b}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.341
}
\keyword{utilities}

