\name{msfd}
\alias{msfd}
\title{ Multiplies a symmetric matrix by a full matrix (double precision) }
\description{
 See Marazzi A. (1993), p.340}
\usage{
msfd(a, b, n)
}
\arguments{
  \item{a}{ See reference}
  \item{b}{ See reference}
  \item{n}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California.v
}
\keyword{utilities}

