\name{mss}
\alias{mss}
\title{ Multiplies a symmetric matrix by a symmetric matrix }
\description{
 See Marazzi A. (1993), p.338}
\usage{
mss(a, b, n)
}
\arguments{
  \item{a}{ See reference}
  \item{b}{ See reference}
  \item{n}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.338
}
\keyword{utilities}

