\name{mssd}
\alias{mssd}
\title{ Multiplies a symmetric matrix by a symmetric matrix (double precision) }
\description{
 See Marazzi A. (1993), p.342}
\usage{
mssd(a, b, n)
}
\arguments{
  \item{a}{ See reference}
  \item{b}{ See reference}
  \item{n}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.342
}
\keyword{utilities}

