\name{mtt1}
\alias{mtt1}
\title{ Multiplies an upper-triangular matrix by its transpose }
\description{
 See Marazzi A. (1993), p.343}
\usage{
mtt1(a, n)
}
\arguments{
  \item{a}{ See reference}
  \item{n}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.343
}
\keyword{utilities}

