\name{mtt2}
\alias{mtt2}
\title{ Multiplies a lower-triangular matrix by its transpose }
\description{
 See Marazzi A. (1993), p.344}
\usage{
mtt2(a, n)
}
\arguments{
  \item{a}{ See reference}
  \item{n}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.344
}
\keyword{utilities}

