\name{mtt3d}
\alias{mtt3d}
\title{ Multiplies a triangular matrix by a triangular matrix (double precision) }
\description{
 See Marazzi A. (1993), p.345}
\usage{
mtt3d(a, b, n)
}
\arguments{
  \item{a}{ See reference}
  \item{b}{ See reference}
  \item{n}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.345
}
\keyword{utilities}

