\name{mtyd}
\alias{mtyd}
\title{ Multiplies an upper-triangular matrix by a vector }
\description{
 See Marazzi A. (1993), p.347}
\usage{
mtyd(a, y, n, iye = 1)
}
\arguments{
  \item{a}{ See reference}
  \item{y}{ See reference}
  \item{n}{ See reference}
  \item{iye}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.347
}
\keyword{utilities}

