\name{myhbhe}
\alias{myhbhe}
\title{ High breakdown point and high efficiency regression with test
                         for bias }
\description{
 See Marazzi A. (1993), p.270}
\usage{
myhbhe(x, y, iseed = .dFvGet()$isd)
}
\arguments{
  \item{x}{ See reference}
  \item{y}{ See reference}
  \item{iseed}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.270
}
\keyword{robust}

