\name{nlgm}
\alias{nlgm}
\title{ Logarithm of the Gamma-function at the point n/2}
\description{
 See Marazzi A. (1993), p.382}
\usage{
nlgm(n)
}
\arguments{
  \item{n}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.382
}
\keyword{utilities}

