\name{nrm2}
\alias{nrm2}
\title{ Forms the Euclidean norm of a vector }
\description{
 See Marazzi A. (1993), p.351}

\usage{
nrm2(x, n = nrow(x), incx = 1)
}
\arguments{
  \item{x}{ See reference}
  \item{n}{ See reference}
  \item{incx}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.351
}
\keyword{utilities}

