\name{nrm2d}
\alias{nrm2d}
\title{ Forms the Euclidean norm of a vector (double precision) }
\description{
 See Marazzi A. (1993), p.351}
\usage{
nrm2d(x, n = nrow(x), incx = 1)
}
\arguments{
  \item{x}{ See reference}
  \item{n}{ See reference}
  \item{incx}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.351
}
\keyword{utilities}

