\name{permc}
\alias{permc}
\title{ Permutes the columns of a matrix by means of transpositions }
\description{
 See Marazzi A. (1993), p.365}
\usage{
permc(x, it, n = nrow(x), iopt = 1)
}
\arguments{
  \item{x}{ See reference}
  \item{it}{ See reference}
  \item{n}{ See reference}
  \item{iopt}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.365
}
\keyword{algebra}
