\name{permv}
\alias{permv}
\title{ Permutes the elements of a vector }
\description{
 See Marazzi A. (1993), p.366}
\usage{
permv(y, it, iopt = 1)
}
\arguments{
  \item{y}{ See reference}
  \item{it}{ See reference}
  \item{iopt}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.366
}
\keyword{algebra}
