\name{poissn}
\alias{poissn}
\title{ Poisson distribution }
\description{
 See Marazzi A. (1993), p.368}
\usage{
poissn(lambda, k)
}
\arguments{
  \item{lambda}{ See reference}
  \item{k}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.368
}
\keyword{distribution}
