\name{precd}
\alias{precd}
\title{Algorithmic determination of the smallest double precision positive 
       number x }
\description{
 See Marazzi A. (1993), p.385}
\usage{
precd()
}

\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.385
}
\keyword{misc}
