\name{probst}
\alias{probst}
\title{ Cumulative t-distribution function }
\description{
 See Marazzi A. (1993), p.377}
\usage{
probst(x, ifn)
}
\arguments{
  \item{x}{ See reference}
  \item{ifn}{ See reference}
}
\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.377
}
\keyword{distribution}
