\name{Psi}
\alias{Psi}
\title{ psi weight function for location and regression }
\description{
 See Marazzi A. (1993), p.319}
\usage{
Psi(svals)
}
\arguments{
  \item{svals}{ A vector of input values }
}
\value{
The values of the psi weight function for each element of svals
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.319
}
\keyword{robust}
