\name{psp}
\alias{psp}
\title{ psi' weight function for location and regression }
\description{
 See Marazzi A. (1993), p.320}
\usage{
psp(s)
}
\arguments{
  \item{s}{ A scalar input value }
}

\value{
The value of the psi' weight function for s
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.320
}
\keyword{robust}
