\name{Psp}
\alias{Psp}
\title{ psi' weight function for location and regression }
\description{
 See Marazzi A. (1993), p.320}
\usage{
Psp(svals)
}
\arguments{
  \item{svals}{ A vector of input values }
}
\value{
The values of the psi' weight function for each element of svals
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.320
}
\keyword{robust}

