\name{QD2coef.f}
\alias{QD2coef.f}
\title{ Auxiliary function to find mu and sigma parameter for extended  loggamma distribution}
\description{
 QD2coef.f computes for a given lambda, the mu and sigma parameters}
\usage{
QD2coef.f(lambda, yc, delta, muI, sigmaI, zero=1e-4)
}
\arguments{
  \item{lambda}{ See reference}
  \item{yc}{ See reference}
  \item{delta}{ See reference}
  \item{muI}{ See reference}
  \item{sigmaI}{ See reference}
  \item{zero}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (2010) Robust estimation of the extended log-gamma (not yet published)
}
\keyword{robust}



