\name{QD2funC.f}
\alias{QD2funC.f}
\title{ Auxiliary function to find lambda parameter for extended  loggamma distribution}
\description{
 QD2funC.f computes a sum of squared residuals for a given lambda}
\usage{
QD2funC.f(lambda, yc, delta, muI, sigmaI, zero=1e-4)
}
\arguments{
  \item{lambda}{ See reference}
  \item{yc}{ See reference}
  \item{delta}{ See reference}
  \item{muI}{ See reference}
  \item{sigmaI}{ See reference}
  \item{zero}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (2010) Robust estimation of the extended log-gamma (not yet published)
}
\keyword{robust}



