\name{Qn.Exp.f}
\alias{Qn.Exp.f}
\title{ Auxiliary function to compute quantiles of survival cdf}
\description{
 Qn.Exp.f computes quantiles of survival cdf}
\usage{
 Qn.Exp.f(p, yc, delta, mu, sigma, lambda, zero=1e-4)
}
\arguments{
  \item{p}{ See reference}
  \item{yc}{ See reference}
  \item{delta}{ See reference}
  \item{mu}{ See reference}
  \item{sigma}{ See reference}
  \item{lambda}{ See reference}
  \item{zero}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (2010) Robust estimation of the extended log-gamma (not yet published)
}
\keyword{robust}



