\name{quant}
\alias{quant}
\title{ Inverse of the standard Gaussian cumulative distribution function}
\description{
 See Marazzi A. (1993), p.372}
\usage{
quant(p)
}
\arguments{
  \item{p}{ See reference}
}

\value{
 See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.372
}
\keyword{distribution}
