\name{Random}
\alias{Random}
\title{ Uniform random number generator }
\description{
 See Marazzi A. (1993), p.386}
\usage{
Random(iseed = .dFvGet()$isd)
}
\arguments{
  \item{iseed}{ See reference}
}
\details{
See reference
}
\value{
See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California.p.386
}
\keyword{distribution}

