\name{Regtau.f}
\alias{Regtau.f}
\title{ Auxiliary function  for the computation of QQopt}
\description{
 QQopt is a resampling algorithm for the determination of the parameters of a Loggamma 
 model}
\usage{
Regtau.f(x, y, b1, c1, b2, c2, N, tol = 1e-6, seed = 567)
}
\arguments{
  \item{x}{ See reference}
  \item{y}{ See reference}
  \item{b1}{ See reference}
  \item{c1}{ See reference}
  \item{b2}{ See reference}
  \item{c2}{ See reference}
  \item{N}{ See reference}
  \item{tol}{ See reference}
  \item{seed}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (2009) Robust estimation of the generalized log-gamma (not yet published)
}
\keyword{robust}


