\name{rho}
\alias{rho}
\title{rho weight function for location and regression }
\description{
 See Marazzi A. (1993), p.320}
\usage{
rho(s)
}
\arguments{
  \item{s}{ A scalar input value }
}
\value{
The value of the rho function for s
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.320
}
\keyword{robust}

