\name{Rho}
\alias{Rho}
\title{ Rho weight function for location and regression }
\description{
  See Marazzi A. (1993), p.320}
\usage{
Rho(svals)
}
\arguments{
  \item{svals}{ A vector of input values}
}
\value{
 The values of the rho function for each element of svals
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.320
}
\keyword{robust}
