\name{ribet0}
\alias{ribet0}
\title{ Computation of the constant Beta0  }
\description{
  See Marazzi A. (1993), p.100}
\usage{
ribet0(wgt, itype = .dFvGet()$ite, isqw = .dFvGet()$isq, 
       tol = .dFvGet()$tlo)
}
\arguments{
  \item{wgt}{ See reference}
  \item{itype}{ See reference}
  \item{isqw}{ See reference}
  \item{tol}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.10
}
\keyword{robust}
