\name{ribeth}
\alias{ribeth}
\title{ Computation of the constant Beta when Chi=Psi.Psi/2 and Psi is the
    Huber function}
\description{
 See Marazzi A. (1993), p.97}
\usage{
ribeth(wgt, d = .dFvGet()$ddd, itype = .dFvGet()$ite)
}
\arguments{
  \item{wgt}{ See reference}
  \item{d}{ See reference}
  \item{itype}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.97
}
\keyword{robust}
