\name{ribetu}
\alias{ribetu}
\title{ Computation of the constant Beta when Chi is a user-supplied function}
\description{
 See Marazzi A. (1993), p.98}
\usage{
ribetu(wgt, exchi = chi, itype = .dFvGet()$ite, upper = .dFvGet()$upr, 
       til = .dFvGet()$tli)
}
\arguments{
  \item{wgt}{ See reference}
  \item{exchi}{ See reference}
  \item{itype}{ See reference}
  \item{upper}{ See reference}
  \item{til}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.98
}
\keyword{robust}
