\name{riclls}
\alias{riclls}
\title{Solution of the least squares problem}
\description{
 See Marazzi A. (1993), p.67}
\usage{
riclls(xt, y, k = np, ix = .dFvGet()$ix1, iy = .dFvGet()$iy1, 
       sf, sg, sh, ip)
}
\arguments{
  \item{xt}{ See reference}
  \item{y}{ See reference}
  \item{k}{ See reference}
  \item{ix}{ See reference}
  \item{iy}{ See reference}
  \item{sf}{ See reference}
  \item{sg}{ See reference}
  \item{sh}{ See reference}
  \item{ip}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.67
}
\keyword{robust}
