\name{rilars}
\alias{rilars}
\title{ Solution of the least absolute residual problem }
\description{
 See Marazzi A. (1993), p.71}
\usage{
rilars(x, y, tol = .dFvGet()$tlu)
}
\arguments{
  \item{x}{ See reference}
  \item{y}{ See reference}
  \item{tol}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.71
}
\keyword{robust}
