\name{ruben}
\alias{ruben}
\title{Cumulative distribution and density function of a linear
       combination of chi-2 random variables}
\description{
 See Marazzi A. (1993), p.375}
\usage{
ruben(xlmbda, delta, mult, x, xmode = 1, maxit = 50, eps = 1e-04)
}
\arguments{
  \item{xlmbda}{ See reference}
  \item{delta}{ See reference}
  \item{mult}{ See reference}
  \item{x}{ See reference}
  \item{xmode}{ See reference}
  \item{maxit}{ See reference}
  \item{eps}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.375
}
\keyword{distribution}
