\name{scald}
\alias{scald}
\title{ Scales a double precision vector by a constant}
\description{
 See Marazzi A. (1993), p.349}
\usage{
scald(x, sa, n = nrow(x), incx = 1)
}
\arguments{
  \item{x}{ See reference}
  \item{sa}{ See reference}
  \item{n}{ See reference}
  \item{incx}{ See reference}
}
\value{
  See reference
}

\seealso{ \code{\link{scal}}}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.349
}
\keyword{algebra}

