\name{srt1}
\alias{srt1}
\title{Sorts the components of a vector in ascending order }
\description{
 See Marazzi A. (1993), p.361}
\usage{
srt1(a, k1 = 1, k2 = n)
}
\arguments{
  \item{a}{ See reference}
  \item{k1}{ See reference}
  \item{k2}{ See reference}
}
\value{
  See reference
}
\seealso{ \code{\link{srt2}}}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.361
}
\keyword{utilities}

