\name{srt2}
\alias{srt2}
\title{Sorts the components of a vector in ascending order and permutes the 
       components of another vector accordingly}
                        
\description{
 See Marazzi A. (1993), p.362}
\usage{
srt2(a, b, k1 = 1, k2 = n)
}
\arguments{
  \item{a}{ See reference}
  \item{b}{ See reference}
  \item{k1}{ See reference}
  \item{k2}{ See reference}
}
\value{
  See reference
}
\seealso{ \code{\link{srt1}}}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.362
}
\keyword{utilities}

