\name{swapd}
\alias{swapd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Interchanges two vectors (double precision)}
\description{
 See Marazzi A. (1993), p.363}
\usage{
swapd(x, y, n = nrow(x), incx = 1, incy = 1)
}
\arguments{
  \item{x}{ See reference}
  \item{y}{ See reference}
  \item{n}{ See reference}
  \item{incx}{ See reference}
  \item{incy}{ See reference}
}
\value{
  See reference
}
\seealso{ \code{\link{swap}} }
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.363
}
\keyword{utilities}

