\name{tauare}
\alias{tauare}
\title{ Asymptotic relative efficiency of the tau-test }
\description{
 See Marazzi A. (1993), p.190}
\usage{
tauare(itype = .dFvGet()$ite, mu, maxit = .dFvGet()$mxe, cpsi, bb, 
       sigmax = 1, upper = .dFvGet()$upr, til = .dFvGet()$tli, 
       tol = .dFvGet()$tlo)
}
\arguments{
  \item{itype}{ See reference}
  \item{mu}{ See reference}
  \item{maxit}{ See reference}
  \item{cpsi}{ See reference}
  \item{bb}{ See reference}
  \item{sigmax}{ See reference}
  \item{upper}{ See reference}
  \item{til}{ See reference}
  \item{tol}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.190
}
\keyword{robust}

