\name{tfrn2t}
\alias{tfrn2t}
\title{Computes the Rn2-test statistic for a linear 
       hypothesis in  canonical form}
\description{
 See Marazzi A. (1993), p.187}
\usage{
tfrn2t(cov, theta, n, nq, tau = .dFvGet()$tua)
}
\arguments{
  \item{cov}{ See reference}
  \item{theta}{ See reference}
  \item{n}{ See reference}
  \item{nq}{ See reference}
  \item{tau}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.187
}
\keyword{robust}

