\name{tftaut}
\alias{tftaut}
\title{ Computes the tau-test statistic for a linear
        hypothesis in canonical form}
\description{
 See Marazzi A. (1993), p.182}
\usage{
tftaut(rs1, rs2, wgt, exrho = rho, np, nq, sigma, itype = .dFvGet()$ite)
}
\arguments{
  \item{rs1}{ See reference}
  \item{rs2}{ See reference}
  \item{wgt}{ See reference}
  \item{exrho}{ See reference}
  \item{np}{ See reference}
  \item{nq}{ See reference}
  \item{sigma}{ See reference}
  \item{itype}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.182
}
\keyword{robust}

