\name{tisrtc}
\alias{tisrtc}
\title{Permutes the columns of the design matrix:
       Predictors in omega are placed in the first q positions }
\description{
 See Marazzi A. (1993), p.188}
\usage{
tisrtc(x, iv, n = nrow(x))
}
\arguments{
  \item{x}{ See reference}
  \item{iv}{ See reference}
  \item{n}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.188
}
\keyword{robust}

