\name{to.double}
\alias{to.double}
\title{ Convert local variable to Fortran double precision }
\description{
  Function only needed for the interface
}
\usage{
to.double(x)
}
\arguments{
  \item{x}{ An R numeric object }
}
\value{
 x converted to double precision
}
\keyword{interface}

