\name{to.single}
\alias{to.single}
\title{ Convert local variable to Fortran single precision }
\description{
  Function only needed for the interface
}
\usage{
to.single(x)
}
\arguments{
  \item{x}{ An R numeric object }
}
\value{
 x converted to single precision
}
\keyword{interface}

