\name{tquant}
\alias{tquant}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Inverse of the cumulative t-distribution function }
\description{
 See Marazzi A. (1993), p.378}
\usage{
tquant(p, ifn)
}
\arguments{
  \item{p}{ See reference}
  \item{ifn}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.378
}
\keyword{distribution}

