\name{ttaskt}
\alias{ttaskt}
\title{ Computes the matrix Ktau }
\description{
 See Marazzi A. (1993), p.184}
\usage{
ttaskt(cov, ainv, np, nq, mdc = np - nq, fact = .dFvGet()$ffc)
}
\arguments{
  \item{cov}{ See reference}
  \item{ainv}{ See reference}
  \item{np}{ See reference}
  \item{nq}{ See reference}
  \item{mdc}{ See reference}
  \item{fact}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.184
}
\keyword{robust}

