\name{tteign}
\alias{tteign}
\title{Computes the eigenvalues of the matrix Ktau }
\description{
 See Marazzi A. (1993), p.186}
\usage{
tteign(covtau, nq, mdc = np - nq)
}
\arguments{
  \item{covtau}{ See reference}
  \item{nq}{ See reference}
  \item{mdc}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.186
}
\keyword{algebra}

