\name{Ucv}
\alias{Ucv}
\title{ u weight function for covariances }
\description{
  See Marazzi A. (1993), p.323}
\usage{
Ucv(svals)
}
\arguments{
  \item{svals}{ A vector of input values }
}
\value{
  The values of the u function for each element of svals
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.323
}
\keyword{robust}

