\name{ugl}
\alias{ugl}
\title{ ub weight function for M-estimates in GLM }
\description{
 See Marazzi A. (1993), p.331}
\usage{
ugl(upar, npar = 4, s)
}
\arguments{
  \item{upar}{ See reference}
  \item{npar}{ See reference}
  \item{s}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.331
}
\keyword{robust}

