\name{upcv}
\alias{upcv}
\title{u' weight function for covariances }
\description{
 See Marazzi A. (1993), p.325}
\usage{
upcv(s)
}
\arguments{
  \item{s}{ A scalar input value}
}
\value{
The value of the u' function for s
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.325
}
\keyword{robust}

