\name{Upcv}
\alias{Upcv}
\title{ u' weight function for covariances}
\description{
 See Marazzi A. (1993), p.325}
\usage{
Upcv(svals)
}
\arguments{
  \item{svals}{ A vector of input values }
}
\value{
The values of the u function for each element of svals
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.325
}
\keyword{robust}

