\name{userfs}
\alias{userfs}
\title{ Dummy u user function }
\description{
See Marazzi A. (1993), p.139}
\usage{
userfs(s)
}
\arguments{
  \item{s}{ A scalar input value }
}
\value{
 The single precision value of the u function for s
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.139
}
\keyword{robust}

