\name{vcv}
\alias{vcv}
\title{ v weight function for covariances }
\description{
 See Marazzi A. (1993) p.327}
\usage{
vcv(s)
}
\arguments{
  \item{s}{ A scalar input value }
}
\value{
The value of the v function for s
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.327
}
\keyword{robust}

