\name{wfshat}
\alias{wfshat}
\title{ Schweppe original weight proposal }
\description{
 See Marazzi A. (1993), p.137}
\usage{
wfshat(xt, n = nrow(xt), sh)
}
\arguments{
  \item{xt}{ See reference}
  \item{n}{ See reference}
  \item{sh}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.137
}
\keyword{robust}

