\name{wimedv}
\alias{wimedv}
\title{ Initial value of the matrix A }
\description{
 See Marazzi A. (1993), p.119}
\usage{
wimedv(x, nobs = nrow(x), itypw = .dFvGet()$itw, 
       init = .dFvGet()$ini, nfirst = nobs)
}
\arguments{
  \item{x}{ See reference}
  \item{nobs}{ See reference}
  \item{itypw}{ See reference}
  \item{init}{ See reference}
  \item{nfirst}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.119
}
\keyword{robust}

