\name{wpcv}
\alias{wpcv}
\title{ w' weight function for covariances  }
\description{
 See Marazzi A. (1993), p.329}
\usage{
wpcv(s)
}
\arguments{
  \item{s}{ A scalar input value }
}
\value{
The value of the w' function for s}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.329
}
\keyword{robust}

