\name{Wpcv}
\alias{Wpcv}
\title{ w' weight function for covariances }
\description{
  See Marazzi A. (1993), p.329}
\usage{
Wpcv(svals)
}
\arguments{
  \item{svals}{A vector of input values }
}
\value{
 The values of the w' function for each element of svals
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.329
}
\keyword{robust}

