\name{www}
\alias{www}
\title{  w weight function }
\description{
 See Marazzi A. (1993), p.330}
\usage{
www(s)
}
\arguments{
  \item{s}{ A scalar input value }
}
\value{
The value of the w weight function for s}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.330
}
\keyword{robust}

