\name{xerf}
\alias{xerf}
\title{Gaussian density function}
\description{
 See Marazzi A. (1993), p.369}
\usage{
xerf(kode = 2, x)
}
\arguments{
  \item{kode}{ See reference}
  \item{x}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.369
}
\keyword{utilities}

