\name{xerp}
\alias{xerp}
\title{ Density of the norm of a standard Gaussian vector with
        p components}
\description{
 See Marazzi A. (1993), p.370}
\usage{
xerp(ip, xlcnst = -1, s)
}
\arguments{
  \item{ip}{ See reference}
  \item{xlcnst}{ See reference}
  \item{s}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.370
}
\keyword{utilities}

