\name{xsy}
\alias{xsy}
\title{ Evaluates a quadratic form }
\description{
 See Marazzi A. (1993), p.352}
\usage{
xsy(x, y, s)
}
\arguments{
  \item{x}{ See reference}
  \item{y}{ See reference}
  \item{s}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.352
}
\keyword{utilities}

