\name{xsyd}
\alias{xsyd}
\title{ Evaluates a quadratic form (double precision) }
\description{
 See Marazzi A. (1993) p.352}
\usage{
xsyd(x, y, s)
}
\arguments{
  \item{x}{ See reference}
  \item{y}{ See reference}
  \item{s}{ See reference}
}

\value{
  See reference
}
\seealso{ \code{\link{xsy}} }
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.352
}
\keyword{utilities}

