\name{zemll}
\alias{zemll}
\title{ Zeng method for censored data}
\description{
 See Reference}
\usage{
zemll(b, x, yo, do)
}
\arguments{
  \item{b}{ See reference}
  \item{x}{ See reference}
  \item{yo}{ See reference}
  \item{do}{ See reference}
}
\value{
  See reference
}
\references{
  Marazzi A. (1993) \emph{Algorithm, Routines, and S functions 
for Robust Statistics}. Wadsworth & Brooks/cole, Pacific Grove,
California. p.216
}
\keyword{Stats}

